requirement.txt
Flask
requests
Pillow
tensorflow
numpy